удалила лишние классы и перенесла классы в java
