+++ src/main/java/pom/MainPage.java
@@ -1,15 +1,12 @@
 package pom;
 
-import org.junit.Assert;
-import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import io.github.bonigarcia.wdm.WebDriverManager;
+
 import java.time.Duration;
-import java.util.ArrayList;
 import java.util.List;
