+++ src/test/java/order/OrderTest.java
@@ -47,7 +47,7 @@ public class OrderTest {
     public static Object[][] getOrderData() {
         return new Object[][]{
                 {"chrome","up","Иван", "Иванов", "ул Елова д 34", "Красносельская","89223451234" ,"06.11.2022","сутки","черный","привезите пораньше"},
-                //{"chrome","down","Петр", "Петров", "ул Родонитовая д 34", "Сокольники","89225678945" ,"17.11.2022","трое суток","серый","спасибо"},
+                {"chrome","down","Петр", "Петров", "ул Родонитовая д 34", "Сокольники","89225678945" ,"17.11.2022","двое суток","серый","спасибо"},
         };
     }
